use updated dataset api for the use cases
rename ellipse to ellipsoid for consistency
