+++ Examples/Python/RunUseCase.py
@@ -23,7 +23,7 @@ if platform.system() == "Darwin":
     binpath = binpath + ":/Applications/ShapeWorks/bin"
 
 parser = argparse.ArgumentParser(description='Example ShapeWorks Pipeline')
-parser.add_argument("--use_case", help="Specify which use case to run, either: ellipse, ellipse_fd, left_atrium, or femur.")
+parser.add_argument("--use_case", help="Specify which use case to run, either: ellipsoid, ellipsoid_fd, left_atrium, or femur.")
 parser.add_argument("--use_subsample", help="Set number of samples to run the pipeline for a subset of data.")
 parser.add_argument("--interactive", help="Run in interactive mode", action="store_true")
 parser.add_argument("--start_with_prepped_data", help="Start with already prepped data", action="store_true")
